move checkTicks into custom_assertions
