+++ test/jasmine/tests/contour_test.js
@@ -10,7 +10,7 @@ var fail = require('../assets/fail_test');
 var createGraphDiv = require('../assets/create_graph_div');
 var destroyGraphDiv = require('../assets/destroy_graph_div');
 var customMatchers = require('../assets/custom_matchers');
-var checkTicks = require('../assets/check_ticks');
+var checkTicks = require('../assets/custom_assertions').checkTicks;
 
 
 describe('contour defaults', function() {
