+++ test/jasmine/tests/plot_api_test.js
@@ -16,7 +16,7 @@ var customMatchers = require('../assets/custom_matchers');
 var createGraphDiv = require('../assets/create_graph_div');
 var destroyGraphDiv = require('../assets/destroy_graph_div');
 var fail = require('../assets/fail_test');
-var checkTicks = require('../assets/check_ticks');
+var checkTicks = require('../assets/custom_assertions').checkTicks;
 var negateIf = require('../assets/negate_if');
